ui: Move ViewRequest out of common
