+++ credentialsd-ui/src/gui/view_model/mod.rs
@@ -13,9 +13,11 @@ use gettextrs::gettext;
 use serde::{Deserialize, Serialize};
 use tracing::{error, info};
 
-use credentialsd_common::model::{Device, Operation, Transport, ViewUpdate};
+use credentialsd_common::model::{Device, Operation, Transport};
 
 use crate::{RequestingApplication, ViewRequest, client::FlowControlClient};
+
+use super::ViewUpdate;
 
 #[derive(Debug)]
 pub(crate) struct ViewModel {
